refactor(worker): split worker.rs into focused modules
Split the 1,834-line worker.rs monolith into five focused modules:

What changed:
- traits.rs (357 lines): Worker trait (40+ methods), ConnectionModeExt
  and WorkerTypeExt extension traits, worker_to_info() helper, and
  re-exports of ConnectionMode/RuntimeType/WorkerType
- basic.rs (1220 lines): BasicWorker struct, Debug impl, Worker impl,
  constants (DEFAULT_BOOTSTRAP_PORT, MOONCAKE_CONNECTOR), WORKER_CLIENT
  static, and all 850 lines of tests (moved from worker.rs inline)
- metadata.rs (138 lines): WorkerMetadata struct with model matching
  methods, WorkerRoutingKeyLoad per-key load tracking
- body.rs (94 lines): WorkerLoadGuard RAII guard, AttachedBody response
  body wrapper
- health_checker.rs (51 lines): HealthChecker handle with graceful
  shutdown (pub(crate))
- Deleted worker.rs, removed #[expect(clippy::module_inception)]
- Updated mod.rs re-exports to route through new module paths
- Updated 10 files with import path changes (worker::worker::X → X
  or worker::traits::X / worker::basic::X / worker::metadata::X)

Why:
worker.rs was a 1,834-line monolith containing the Worker trait,
BasicWorker impl, HealthChecker, WorkerMetadata, WorkerRoutingKeyLoad,
WorkerLoadGuard, AttachedBody, constants, and 850 lines of tests.
This split is prerequisite for the worker module deep refactor — later
PRs need to modify Worker trait, BasicWorker, and HealthChecker
independently.

How:
Pure code extraction — no behavioral changes. Each section moved to
its natural module. Tests moved to basic.rs with updated imports.
active_routing_keys field widened from private to pub(crate) so tests
in basic.rs can access WorkerRoutingKeyLoad internals. All 133 worker
unit tests and 100 integration tests pass identically.

Refs: worker-module-deep-refactor plan (PR 3)
