+++ model_gateway/src/routers/openai/realtime/rest.rs
@@ -11,7 +11,7 @@ use tracing::error;
 use crate::{
     observability::metrics::{metrics_labels, Metrics},
     routers::header_utils::extract_auth_header,
-    worker::{worker::WorkerLoadGuard, Worker},
+    worker::{Worker, WorkerLoadGuard},
 };
 
 /// Forward a realtime REST request to the upstream worker.
