+++ model_gateway/src/worker/basic.rs
@@ -9,17 +9,17 @@ use std::{
 
 use arc_swap::ArcSwap;
 use async_trait::async_trait;
-use axum::body::Body;
-// Re-export protocol types as the canonical types for the gateway
-pub use openai_protocol::worker::{ConnectionMode, RuntimeType, WorkerType};
 use openai_protocol::{
     model_card::ModelCard,
-    model_type::{Endpoint, ModelType},
-    worker::{HealthCheckConfig, ProviderType, WorkerInfo, WorkerModels, WorkerSpec, WorkerStatus},
+    worker::{ConnectionMode, WorkerModels, WorkerType},
 };
 use tokio::{sync::OnceCell, time};
 
-use super::{CircuitBreaker, ResolvedResilience, WorkerError, WorkerResult, UNKNOWN_MODEL_ID};
+use super::{
+    metadata::{WorkerMetadata, WorkerRoutingKeyLoad},
+    traits::{Worker, WorkerTypeExt},
+    CircuitBreaker, ResolvedResilience, WorkerError, WorkerResult,
+};
 use crate::{
     observability::metrics::{metrics_labels, Metrics},
     routers::grpc::client::GrpcClient,
@@ -44,454 +44,6 @@ static WORKER_CLIENT: LazyLock<reqwest::Client> = LazyLock::new(|| {
         .build()
         .expect("Failed to create worker HTTP client")
 });
-
-pub struct WorkerRoutingKeyLoad {
-    url: String,
-    active_routing_keys: dashmap::DashMap<String, usize>,
-}
-
-impl WorkerRoutingKeyLoad {
-    pub fn new(url: impl Into<String>) -> Self {
-        Self {
-            url: url.into(),
-            active_routing_keys: dashmap::DashMap::new(),
-        }
-    }
-
-    pub fn value(&self) -> usize {
-        self.active_routing_keys.len()
-    }
-
-    pub fn increment(&self, routing_key: &str) {
-        *self
-            .active_routing_keys
-            .entry(routing_key.to_string())
-            .or_insert(0) += 1;
-        self.update_metrics();
-    }
-
-    pub fn decrement(&self, routing_key: &str) {
-        use dashmap::mapref::entry::Entry;
-
-        match self.active_routing_keys.entry(routing_key.to_string()) {
-            Entry::Occupied(mut entry) => {
-                let counter = entry.get_mut();
-                if *counter > 0 {
-                    *counter -= 1;
-                    if *counter == 0 {
-                        entry.remove();
-                    }
-                } else {
-                    tracing::warn!(
-                        worker_url = %self.url,
-                        routing_key = %routing_key,
-                        "Attempted to decrement routing key counter that is already at 0"
-                    );
-                }
-            }
-            Entry::Vacant(_) => {
-                tracing::warn!(
-                    worker_url = %self.url,
-                    routing_key = %routing_key,
-                    "Attempted to decrement non-existent routing key"
-                );
-            }
-        }
-        self.update_metrics();
-    }
-
-    fn update_metrics(&self) {
-        Metrics::set_worker_routing_keys_active(&self.url, self.value());
-    }
-}
-
-impl fmt::Debug for WorkerRoutingKeyLoad {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("WorkerRoutingKeyLoad")
-            .field("url", &self.url)
-            .field("active_routing_keys", &self.value())
-            .finish()
-    }
-}
-
-/// Core worker abstraction that represents a backend service
-#[async_trait]
-pub trait Worker: Send + Sync + fmt::Debug {
-    /// Get the worker's URL
-    fn url(&self) -> &str;
-    /// Get the worker's API key
-    fn api_key(&self) -> Option<&String>;
-    /// Get the worker's type (Regular, Prefill, or Decode)
-    /// Returns a reference to avoid cloning on every access
-    fn worker_type(&self) -> &WorkerType;
-
-    /// Get the worker's connection mode (HTTP or gRPC)
-    /// Returns a reference to avoid cloning on every access
-    fn connection_mode(&self) -> &ConnectionMode;
-
-    /// Get the bootstrap hostname for PD mode
-    /// Returns cached hostname parsed from URL at construction time
-    fn bootstrap_host(&self) -> &str {
-        &self.metadata().spec.bootstrap_host
-    }
-
-    /// Get the bootstrap port for PD mode
-    /// Returns cached port from WorkerType::Prefill
-    fn bootstrap_port(&self) -> Option<u16> {
-        self.metadata().spec.bootstrap_port
-    }
-
-    /// Check if the worker is currently healthy
-    fn is_healthy(&self) -> bool;
-
-    /// Set the worker's health status
-    fn set_healthy(&self, healthy: bool);
-
-    /// Perform an async health check on the worker
-    async fn check_health_async(&self) -> WorkerResult<()>;
-
-    /// Get the current load (number of active requests)
-    fn load(&self) -> usize;
-
-    /// Increment the load counter
-    fn increment_load(&self);
-
-    /// Decrement the load counter
-    fn decrement_load(&self);
-
-    /// Reset the load counter to 0 (for sync/recovery)
-    fn reset_load(&self) {}
-
-    /// Get the worker routing key load tracker
-    fn worker_routing_key_load(&self) -> &WorkerRoutingKeyLoad;
-
-    /// Get the number of processed requests
-    fn processed_requests(&self) -> usize;
-
-    /// Increment the processed requests counter
-    fn increment_processed(&self);
-
-    /// Get worker-specific metadata
-    fn metadata(&self) -> &WorkerMetadata;
-
-    /// Get the circuit breaker for this worker.
-    ///
-    /// **Do not call from routers.** Use `record_outcome(status_code)` to
-    /// record request outcomes and `is_available()` to check worker health.
-    fn circuit_breaker(&self) -> &CircuitBreaker;
-
-    /// Check if the worker is available (healthy + circuit closed/half-open)
-    fn is_available(&self) -> bool {
-        self.is_healthy() && self.circuit_breaker().can_execute()
-    }
-
-    /// Record the outcome of a request based on the HTTP status code.
-    ///
-    /// The worker decides whether the status is a CB failure using its
-    /// per-worker `retryable_status_codes` set (default: 408, 429, 5xx).
-    /// Callers just pass the status — no need to interpret it.
-    ///
-    /// For transport/connection errors where no HTTP response is received,
-    /// pass the status code returned to the client (e.g., 502 for a send
-    /// error, 504 for a timeout).
-    fn record_outcome(&self, status_code: u16) {
-        let is_failure = self
-            .resilience()
-            .retryable_status_codes
-            .contains(&status_code);
-        self.circuit_breaker().record_outcome(!is_failure);
-    }
-
-    /// Get the resolved resilience config for this worker.
-    fn resilience(&self) -> &ResolvedResilience;
-
-    /// Get the per-worker HTTP client.
-    fn http_client(&self) -> &reqwest::Client;
-
-    /// Check if this worker is DP-aware
-    fn is_dp_aware(&self) -> bool {
-        self.metadata().spec.dp_rank.is_some()
-    }
-
-    /// Get the base URL without any DP rank suffix
-    fn base_url(&self) -> &str {
-        self.metadata()
-            .spec
-            .dp_base_url
-            .as_deref()
-            .unwrap_or_else(|| self.url())
-    }
-
-    /// Get DP rank if this is a DP-aware worker
-    fn dp_rank(&self) -> Option<usize> {
-        self.metadata().spec.dp_rank
-    }
-
-    /// Get DP size if this worker is part of a DP group
-    fn dp_size(&self) -> Option<usize> {
-        self.metadata().spec.dp_size
-    }
-
-    /// Transform a request for DP-aware routing
-    async fn prepare_request(&self, mut req: serde_json::Value) -> WorkerResult<serde_json::Value> {
-        if let Some(rank) = self.metadata().spec.dp_rank {
-            if let Some(map) = req.as_object_mut() {
-                map.insert("data_parallel_rank".to_string(), serde_json::json!(rank));
-                Ok(req)
-            } else {
-                Err(WorkerError::InvalidConfiguration {
-                    message: "Request must be a JSON object for DP-aware routing".to_string(),
-                })
-            }
-        } else {
-            Ok(req)
-        }
-    }
-
-    /// Get the actual endpoint URL for requests
-    fn endpoint_url(&self, route: &str) -> String {
-        format!("{}{}", self.base_url(), route)
-    }
-
-    /// Check if this worker can handle a specific request
-    fn can_handle(&self, _req: &serde_json::Value) -> bool {
-        true
-    }
-
-    /// Get the model ID this worker serves
-    /// Checks ModelCards first, then falls back to labels
-    fn model_id(&self) -> &str {
-        // Check ModelCards first
-        self.metadata()
-            .spec
-            .models
-            .primary()
-            .map(|m| m.id.as_str())
-            .or_else(|| {
-                // Fall back to labels
-                self.metadata()
-                    .spec
-                    .labels
-                    .get("model_id")
-                    .map(|s| s.as_str())
-            })
-            .unwrap_or(UNKNOWN_MODEL_ID)
-    }
-
-    /// Get the priority of this worker (higher value = higher priority)
-    fn priority(&self) -> u32 {
-        self.metadata().spec.priority
-    }
-
-    /// Get the cost factor of this worker (baseline = 1.0)
-    fn cost(&self) -> f32 {
-        self.metadata().spec.cost
-    }
-
-    /// Get tokenizer path for a specific model.
-    fn tokenizer_path(&self, model_id: &str) -> Option<&str> {
-        self.metadata()
-            .find_model(model_id)
-            .and_then(|m| m.tokenizer_path.as_deref())
-    }
-
-    /// Get reasoning parser for a specific model.
-    fn reasoning_parser(&self, model_id: &str) -> Option<&str> {
-        self.metadata()
-            .find_model(model_id)
-            .and_then(|m| m.reasoning_parser.as_deref())
-    }
-
-    /// Get tool parser for a specific model.
-    fn tool_parser(&self, model_id: &str) -> Option<&str> {
-        self.metadata()
-            .find_model(model_id)
-            .and_then(|m| m.tool_parser.as_deref())
-    }
-
-    /// Get chat template for a specific model.
-    fn chat_template(&self, model_id: &str) -> Option<&str> {
-        self.metadata()
-            .find_model(model_id)
-            .and_then(|m| m.chat_template.as_deref())
-    }
-
-    /// Get the default provider type for this worker.
-    /// `None` means native/passthrough.
-    fn default_provider(&self) -> Option<&ProviderType> {
-        self.metadata().spec.provider.as_ref()
-    }
-
-    /// Get provider for a specific model.
-    /// Priority: ModelCard.provider > worker.default_provider
-    fn provider_for_model(&self, model_id: &str) -> Option<&ProviderType> {
-        self.metadata().provider_for_model(model_id)
-    }
-
-    /// Check if a model is a classifier (has id2label mapping).
-    fn is_classifier(&self, model_id: &str) -> bool {
-        self.metadata()
-            .find_model(model_id)
-            .map(|m| m.is_classifier())
-            .unwrap_or(false)
-    }
-
-    /// Get the id2label mapping for a classification model.
-    /// Returns None if model is not a classifier or not found.
-    fn id2label(&self, model_id: &str) -> Option<&std::collections::HashMap<u32, String>> {
-        self.metadata()
-            .find_model(model_id)
-            .filter(|m| m.is_classifier())
-            .map(|m| &m.id2label)
-    }
-
-    /// Get the number of classification labels for a model.
-    fn num_labels(&self, model_id: &str) -> u32 {
-        self.metadata()
-            .find_model(model_id)
-            .map(|m| m.num_labels)
-            .unwrap_or(0)
-    }
-
-    /// Get label for a class index from a classification model.
-    /// Returns generic label (LABEL_N) if model not found or index not in mapping.
-    fn get_label(&self, model_id: &str, class_idx: u32) -> String {
-        self.metadata()
-            .find_model(model_id)
-            .map(|m| m.get_label(class_idx))
-            .unwrap_or_else(|| format!("LABEL_{class_idx}"))
-    }
-
-    /// Check if this worker supports a specific model.
-    /// If models list is empty, worker accepts any model.
-    fn supports_model(&self, model_id: &str) -> bool {
-        self.metadata().supports_model(model_id)
-    }
-
-    /// Check if this worker supports an endpoint for a given model.
-    /// Falls back to default_model_type if model not found.
-    fn supports_endpoint(&self, model_id: &str, endpoint: Endpoint) -> bool {
-        self.metadata().supports_endpoint(model_id, endpoint)
-    }
-
-    /// Get all models this worker can serve.
-    fn models(&self) -> Vec<ModelCard> {
-        self.metadata().spec.models.all().to_vec()
-    }
-
-    /// Set models for this worker (for lazy discovery).
-    /// Default implementation does nothing - only BasicWorker supports this.
-    fn set_models(&self, _models: Vec<ModelCard>) {
-        // Default: no-op. BasicWorker overrides this.
-    }
-
-    /// Check if models have been discovered for this worker.
-    /// Returns true if models were set via set_models() or if metadata has models.
-    fn has_models_discovered(&self) -> bool {
-        !self.metadata().spec.models.is_wildcard()
-    }
-
-    /// Get or create a gRPC client for this worker
-    /// Returns None for HTTP workers, Some(client) for gRPC workers
-    async fn get_grpc_client(&self) -> WorkerResult<Option<Arc<GrpcClient>>>;
-
-    /// Reset the gRPC client connection (for reconnection scenarios)
-    /// No-op for HTTP workers
-    async fn reset_grpc_client(&self) -> WorkerResult<()> {
-        Ok(())
-    }
-    async fn grpc_health_check(&self) -> WorkerResult<bool>;
-    async fn http_health_check(&self) -> WorkerResult<bool>;
-}
-
-/// Extension trait for model_gateway-specific ConnectionMode methods.
-pub(crate) trait ConnectionModeExt {
-    fn as_metric_label(&self) -> &'static str;
-}
-
-impl ConnectionModeExt for ConnectionMode {
-    fn as_metric_label(&self) -> &'static str {
-        match self {
-            ConnectionMode::Http => metrics_labels::CONNECTION_HTTP,
-            ConnectionMode::Grpc => metrics_labels::CONNECTION_GRPC,
-        }
-    }
-}
-
-/// Extension trait for model_gateway-specific WorkerType methods.
-pub(crate) trait WorkerTypeExt {
-    fn as_metric_label(&self) -> &'static str;
-}
-
-impl WorkerTypeExt for WorkerType {
-    fn as_metric_label(&self) -> &'static str {
-        match self {
-            WorkerType::Regular => metrics_labels::WORKER_REGULAR,
-            WorkerType::Prefill => metrics_labels::WORKER_PREFILL,
-            WorkerType::Decode => metrics_labels::WORKER_DECODE,
-        }
-    }
-}
-
-/// Metadata associated with a worker.
-///
-/// Embeds [`WorkerSpec`] for identity/config fields shared with the
-/// protocol layer, plus internal-only fields for health checking and
-/// endpoint routing.
-#[derive(Debug, Clone)]
-pub struct WorkerMetadata {
-    /// Protocol-level worker identity and configuration.
-    pub spec: WorkerSpec,
-    /// Resolved health check config (router defaults + per-worker overrides).
-    /// This is the concrete config used at runtime; `spec.health` only stores
-    /// the partial overrides from the API layer.
-    pub health_config: HealthCheckConfig,
-    /// Health check endpoint path (internal-only, from router config).
-    pub health_endpoint: String,
-}
-
-impl WorkerMetadata {
-    /// Find a model card by ID (including aliases)
-    pub fn find_model(&self, model_id: &str) -> Option<&ModelCard> {
-        self.spec.models.find(model_id)
-    }
-
-    /// Check if this worker can serve a given model.
-    /// Wildcard workers accept any model.
-    pub fn supports_model(&self, model_id: &str) -> bool {
-        self.spec.models.supports(model_id)
-    }
-
-    /// Check if this worker supports an endpoint for a given model.
-    /// Falls back to LLM capabilities if model not found — this is safe because
-    /// non-LLM workers (embeddings, rerank) are always registered with explicit
-    /// models via discovery, never as wildcards.
-    pub fn supports_endpoint(&self, model_id: &str, endpoint: Endpoint) -> bool {
-        if let Some(model) = self.find_model(model_id) {
-            model.supports_endpoint(endpoint)
-        } else {
-            ModelType::LLM.supports_endpoint(endpoint)
-        }
-    }
-
-    /// Get the provider for a given model.
-    /// Returns the model's provider if found, otherwise the worker's default provider.
-    pub fn provider_for_model(&self, model_id: &str) -> Option<&ProviderType> {
-        self.find_model(model_id)
-            .and_then(|m| m.provider.as_ref())
-            .or(self.spec.provider.as_ref())
-    }
-
-    /// Get all model IDs this worker can serve
-    pub fn model_ids(&self) -> impl Iterator<Item = &str> {
-        self.spec.models.iter().map(|m| m.id.as_str())
-    }
-
-    /// Check if this worker is in wildcard mode (accepts any model).
-    pub fn is_wildcard(&self) -> bool {
-        self.spec.models.is_wildcard()
-    }
-}
 
 /// Basic worker implementation
 #[derive(Clone)]
@@ -822,173 +374,23 @@ impl Worker for BasicWorker {
     }
 }
 
-/// RAII guard for worker load management
-///
-/// Automatically decrements worker load when dropped. Can be attached to
-/// an axum Response to tie the guard's lifetime to the response body,
-/// which is essential for streaming responses where the function returns
-/// immediately but the stream continues in the background.
-pub struct WorkerLoadGuard {
-    worker: Arc<dyn Worker>,
-    routing_key: Option<String>,
-}
-
-impl WorkerLoadGuard {
-    pub fn new(worker: Arc<dyn Worker>, headers: Option<&http::HeaderMap>) -> Self {
-        use crate::routers::header_utils::extract_routing_key;
-
-        worker.increment_load();
-
-        let routing_key = extract_routing_key(headers).map(String::from);
-
-        if let Some(ref key) = routing_key {
-            worker.worker_routing_key_load().increment(key);
-        }
-
-        Self {
-            worker,
-            routing_key,
-        }
-    }
-}
-
-impl Drop for WorkerLoadGuard {
-    fn drop(&mut self) {
-        self.worker.decrement_load();
-        if let Some(ref key) = self.routing_key {
-            self.worker.worker_routing_key_load().decrement(key);
-        }
-    }
-}
-
-/// Body wrapper that holds an attached value.
-///
-/// When this body is dropped (stream ends or client disconnects),
-/// the attached value is dropped automatically. This is useful for RAII guards
-/// like WorkerLoadGuard that need to be tied to a response body's lifetime.
-pub struct AttachedBody<T> {
-    inner: Body,
-    _attached: T,
-}
-
-impl<T> AttachedBody<T> {
-    pub fn new(inner: Body, attached: T) -> Self {
-        Self {
-            inner,
-            _attached: attached,
-        }
-    }
-}
-
-impl<T: Send + Unpin + 'static> AttachedBody<T> {
-    pub fn wrap_response(
-        response: axum::response::Response,
-        attached: T,
-    ) -> axum::response::Response {
-        let (parts, body) = response.into_parts();
-        axum::response::Response::from_parts(parts, Body::new(Self::new(body, attached)))
-    }
-}
-
-impl<T: Send + Unpin + 'static> http_body::Body for AttachedBody<T> {
-    type Data = bytes::Bytes;
-    type Error = axum::Error;
-
-    fn poll_frame(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Option<Result<http_body::Frame<Self::Data>, Self::Error>>> {
-        let this = self.get_mut();
-        std::pin::Pin::new(&mut this.inner).poll_frame(cx)
-    }
-
-    fn is_end_stream(&self) -> bool {
-        self.inner.is_end_stream()
-    }
-
-    fn size_hint(&self) -> http_body::SizeHint {
-        self.inner.size_hint()
-    }
-}
-
-/// Health checker handle with graceful shutdown.
-///
-/// The checker sleeps until the next worker is due for a health check,
-/// so it wakes only when there is actual work to do.
-pub(crate) struct HealthChecker {
-    handle: Option<tokio::task::JoinHandle<()>>,
-    shutdown_notify: Arc<tokio::sync::Notify>,
-}
-
-impl fmt::Debug for HealthChecker {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("HealthChecker").finish()
-    }
-}
-
-impl HealthChecker {
-    pub fn new(
-        handle: tokio::task::JoinHandle<()>,
-        shutdown_notify: Arc<tokio::sync::Notify>,
-    ) -> Self {
-        Self {
-            handle: Some(handle),
-            shutdown_notify,
-        }
-    }
-
-    /// Shutdown the health checker gracefully.
-    /// Wakes the sleeping task immediately so it can exit cleanly.
-    /// Prefer this over dropping when you can `.await` — it lets the
-    /// current health-check iteration finish instead of aborting mid-flight.
-    #[expect(
-        dead_code,
-        reason = "Drop::drop handles abort; this exists for graceful shutdown when an async context is available"
-    )]
-    pub async fn shutdown(&mut self) {
-        self.shutdown_notify.notify_one();
-        if let Some(handle) = self.handle.take() {
-            let _ = handle.await;
-        }
-    }
-}
-
-impl Drop for HealthChecker {
-    fn drop(&mut self) {
-        if let Some(handle) = self.handle.take() {
-            handle.abort();
-        }
-    }
-}
-
-/// Helper to convert Worker trait object to WorkerInfo struct
-pub fn worker_to_info(worker: &Arc<dyn Worker>) -> WorkerInfo {
-    let metadata = worker.metadata();
-    let spec = metadata.spec.clone();
-    let is_healthy = worker.is_healthy();
-
-    WorkerInfo {
-        id: worker.url().to_string(),
-        model_id: spec.models.primary().map(|m| m.id.clone()),
-        spec,
-        is_healthy,
-        status: Some(if is_healthy {
-            WorkerStatus::Ready
-        } else {
-            WorkerStatus::NotReady
-        }),
-        load: worker.load(),
-        job_status: None,
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use std::{thread, time::Duration};
+    use std::{sync::Arc, thread, time::Duration};
+
+    use openai_protocol::{
+        model_card::ModelCard,
+        worker::{HealthCheckConfig, WorkerModels, WorkerSpec},
+    };
+    use tokio::time;
 
     use super::*;
     use crate::worker::{
+        body::WorkerLoadGuard,
         circuit_breaker::{CircuitBreakerConfig, CircuitState},
+        error::WorkerError,
+        metadata::{WorkerMetadata, WorkerRoutingKeyLoad},
+        traits::{Worker, WorkerType},
         BasicWorkerBuilder,
     };
 
@@ -1015,7 +417,6 @@ mod tests {
 
     #[test]
     fn test_health_config_default() {
-        use openai_protocol::worker::HealthCheckConfig;
         let config = HealthCheckConfig::default();
         assert_eq!(config.timeout_secs, 30);
         assert_eq!(config.check_interval_secs, 60);
@@ -1026,7 +427,6 @@ mod tests {
 
     #[test]
     fn test_health_config_custom() {
-        use openai_protocol::worker::HealthCheckConfig;
         let config = HealthCheckConfig {
             timeout_secs: 10,
             check_interval_secs: 60,
@@ -1043,7 +443,6 @@ mod tests {
 
     #[test]
     fn test_basic_worker_creation() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .build();
@@ -1060,7 +459,6 @@ mod tests {
         labels.insert("env".to_string(), "prod".to_string());
         labels.insert("zone".to_string(), "us-west".to_string());
 
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .labels(labels.clone())
@@ -1071,7 +469,6 @@ mod tests {
 
     #[test]
     fn test_worker_with_health_config() {
-        use openai_protocol::worker::HealthCheckConfig;
         let custom_config = HealthCheckConfig {
             timeout_secs: 15,
             check_interval_secs: 45,
@@ -1080,7 +477,6 @@ mod tests {
             disable_health_check: false,
         };
 
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .health_config(custom_config.clone())
@@ -1094,7 +490,6 @@ mod tests {
 
     #[test]
     fn test_worker_url() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://worker1:8080")
             .worker_type(WorkerType::Regular)
             .build();
@@ -1103,7 +498,6 @@ mod tests {
 
     #[test]
     fn test_worker_type_getter() {
-        use crate::worker::BasicWorkerBuilder;
         let regular = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .build();
@@ -1124,7 +518,6 @@ mod tests {
 
     #[test]
     fn test_health_status() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .build();
@@ -1140,7 +533,6 @@ mod tests {
 
     #[test]
     fn test_load_counter_operations() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .build();
@@ -1167,7 +559,6 @@ mod tests {
 
     #[test]
     fn test_processed_counter() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .build();
@@ -1182,7 +573,6 @@ mod tests {
 
     #[tokio::test]
     async fn test_concurrent_load_increments() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = Arc::new(
             BasicWorkerBuilder::new("http://test:8080")
                 .worker_type(WorkerType::Regular)
@@ -1212,7 +602,6 @@ mod tests {
 
     #[tokio::test]
     async fn test_concurrent_load_decrements() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = Arc::new(
             BasicWorkerBuilder::new("http://test:8080")
                 .worker_type(WorkerType::Regular)
@@ -1247,7 +636,6 @@ mod tests {
 
     #[tokio::test]
     async fn test_concurrent_health_updates() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = Arc::new(
             BasicWorkerBuilder::new("http://test:8080")
                 .worker_type(WorkerType::Regular)
@@ -1276,7 +664,6 @@ mod tests {
 
     #[test]
     fn test_create_regular_worker() {
-        use crate::worker::BasicWorkerBuilder;
         let worker: Box<dyn Worker> = Box::new(
             BasicWorkerBuilder::new("http://regular:8080")
                 .worker_type(WorkerType::Regular)
@@ -1288,7 +675,6 @@ mod tests {
 
     #[test]
     fn test_create_prefill_worker() {
-        use crate::worker::BasicWorkerBuilder;
         let worker1: Box<dyn Worker> = Box::new(
             BasicWorkerBuilder::new("http://prefill:8080")
                 .worker_type(WorkerType::Prefill)
@@ -1310,7 +696,6 @@ mod tests {
 
     #[test]
     fn test_create_decode_worker() {
-        use crate::worker::BasicWorkerBuilder;
         let worker: Box<dyn Worker> = Box::new(
             BasicWorkerBuilder::new("http://decode:8080")
                 .worker_type(WorkerType::Decode)
@@ -1322,7 +707,6 @@ mod tests {
 
     #[tokio::test]
     async fn test_check_health_async() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .build();
@@ -1336,8 +720,6 @@ mod tests {
     #[expect(clippy::print_stderr)]
     fn test_load_counter_performance() {
         use std::time::Instant;
-
-        use crate::worker::BasicWorkerBuilder;
 
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
@@ -1475,7 +857,6 @@ mod tests {
 
     #[test]
     fn test_worker_circuit_breaker() {
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .build();
@@ -1506,7 +887,6 @@ mod tests {
             window_duration: Duration::from_secs(60),
         };
 
-        use crate::worker::BasicWorkerBuilder;
         let worker = BasicWorkerBuilder::new("http://test:8080")
             .worker_type(WorkerType::Regular)
             .circuit_breaker_config(config)
@@ -1545,7 +925,6 @@ mod tests {
 
     #[tokio::test]
     async fn test_mixed_worker_types() {
-        use crate::worker::BasicWorkerBuilder;
         let regular: Box<dyn Worker> = Box::new(
             BasicWorkerBuilder::new("http://regular:8080")
                 .worker_type(WorkerType::Regular)
@@ -1614,8 +993,6 @@ mod tests {
         assert_eq!(workers[5].worker_type(), &WorkerType::Decode);
     }
 
-    // === Phase 1.3: WorkerMetadata model methods tests ===
-
     #[test]
     fn test_worker_metadata_empty_models_accepts_all() {
         let metadata = WorkerMetadata {
@@ -1624,7 +1001,6 @@ mod tests {
             health_endpoint: "/health".to_string(),
         };
 
-        // Empty models list should accept any model
         assert!(metadata.supports_model("any-model"));
         assert!(metadata.supports_model("gpt-4"));
         assert!(metadata.supports_model("llama-3.1"));
@@ -1632,8 +1008,6 @@ mod tests {
 
     #[test]
     fn test_worker_metadata_find_model() {
-        use super::ModelCard;
-
         let model1 = ModelCard::new("meta-llama/Llama-3.1-8B")
             .with_alias("llama-3.1-8b")
             .with_alias("llama3.1");
@@ -1647,15 +1021,10 @@ mod tests {
             health_endpoint: "/health".to_string(),
         };
 
-        // Find by primary ID
         assert!(metadata.find_model("meta-llama/Llama-3.1-8B").is_some());
         assert!(metadata.find_model("gpt-4o").is_some());
-
-        // Find by alias
         assert!(metadata.find_model("llama-3.1-8b").is_some());
         assert!(metadata.find_model("llama3.1").is_some());
-
-        // Not found
         assert!(metadata.find_model("unknown-model").is_none());
     }
 
@@ -1731,8 +1100,6 @@ mod tests {
 
     #[test]
     fn test_worker_load_guard_with_routing_key() {
-        use crate::worker::BasicWorkerBuilder;
-
         let worker: Arc<dyn Worker> = Arc::new(
             BasicWorkerBuilder::new("http://test:8000")
                 .worker_type(WorkerType::Regular)
@@ -1757,8 +1124,6 @@ mod tests {
 
     #[test]
     fn test_worker_load_guard_without_routing_key() {
-        use crate::worker::BasicWorkerBuilder;
-
         let worker: Arc<dyn Worker> = Arc::new(
             BasicWorkerBuilder::new("http://test:8000")
                 .worker_type(WorkerType::Regular)
@@ -1780,8 +1145,6 @@ mod tests {
 
     #[test]
     fn test_worker_load_guard_multiple_same_routing_key() {
-        use crate::worker::BasicWorkerBuilder;
-
         let worker: Arc<dyn Worker> = Arc::new(
             BasicWorkerBuilder::new("http://test:8000")
                 .worker_type(WorkerType::Regular)
@@ -1812,12 +1175,10 @@ mod tests {
     fn test_lazy_discovered_models_override_wildcard() {
         let worker = BasicWorkerBuilder::new("http://test:8080").build();
 
-        // Wildcard worker starts with no models listed, but accepts any model
         assert!(worker.models().is_empty());
         assert!(!worker.has_models_discovered());
-        assert!(worker.supports_model("gpt-4o-mini")); // wildcard accepts anything
+        assert!(worker.supports_model("gpt-4o-mini"));
 
-        // Simulate lazy discovery via set_models
         let discovered = vec![
             ModelCard::new("gpt-4o-mini"),
             ModelCard::new("text-embedding-3-small"),
