+++ model_gateway/src/workflow/steps/local/create_worker.rs
@@ -10,8 +10,8 @@ use wfaas::{StepExecutor, StepId, StepResult, WorkflowContext, WorkflowError, Wo
 use crate::{
     worker::{
         circuit_breaker::CircuitBreakerConfig, http_client::build_worker_http_client,
-        resilience::resolve_resilience, worker::RuntimeType, BasicWorkerBuilder, ConnectionMode,
-        Worker, UNKNOWN_MODEL_ID,
+        resilience::resolve_resilience, BasicWorkerBuilder, ConnectionMode, RuntimeType, Worker,
+        UNKNOWN_MODEL_ID,
     },
     workflow::data::{WorkerKind, WorkerWorkflowData},
 };
